v8 candidate
git-svn-id: https://nlg0.isi.edu/svn/sbmt/trunk/graehl@1187 7f356e55-ce0a-0410-be72-bc6dfe96aa1c

+++ shared/memory_stats.hpp
@@ -47,18 +47,18 @@ struct memory_stats  {
     
     size_type program_allocated() const 
     {
-        return size_type(info.uordblks);
+        return size_type((unsigned)info.uordblks);
     }
 
     // may only grown monotonically (may not reflect free())
     size_type system_allocated() const 
     {
-        return size_type(info.arena);
+        return size_type((unsigned)info.arena);
     }
     
     size_type memory_mapped() const
     {
-        return size_type(info.hblkhd);
+        return size_type((unsigned)info.hblkhd);
     }
 };
